feat: hot reloadable config
