+++ packages/config/package.json
@@ -15,6 +15,7 @@
     "prettier:fix": "prettier --write \"**/*.{css,html,js,jsx,yml}\""
   },
   "dependencies": {
+    "@rm/logger": "*",
     "config": "^3.3.12"
   }
 }
