+++ packages/logger/package.json
@@ -12,7 +12,6 @@
     "prettier:fix": "prettier --write \"**/*.{css,html,js,jsx,yml}\""
   },
   "dependencies": {
-    "@rm/config": "*",
     "chalk": "^4",
     "loglevel": "^1.8.1"
   }
