+++ server/src/routes/api/v1/config.js
@@ -1,6 +1,9 @@
 // @ts-check
 const router = require('express').Router()
 const { log, HELPERS } = require('@rm/logger')
+const state = require('../../../services/state')
+const { bindConnections } = require('../../../models')
+const { loadLatestAreas } = require('../../../services/areas')
 
 router.get('/', (req, res) => {
   const config = require('@rm/config')
@@ -25,6 +28,26 @@ router.get('/', (req, res) => {
           : [],
       },
     })
+  } catch (e) {
+    log.error(HELPERS.api, req.originalUrl, e)
+    res.status(500).json({ status: 'error', reason: e.message })
+  }
+})
+
+router.get('/reload', async (req, res) => {
+  try {
+    const newConfig = require('@rm/config').reload()
+
+    const newState = state.reload()
+
+    bindConnections(newState.db)
+
+    await newState.db.getDbContext()
+    await newState.loadLocalContexts()
+    await newState.loadExternalContexts()
+    newConfig.setAreas(await loadLatestAreas())
+
+    res.status(200).json({ status: 'ok' })
   } catch (e) {
     log.error(HELPERS.api, req.originalUrl, e)
     res.status(500).json({ status: 'error', reason: e.message })
