+++ server/src/services/filters/pokemon/constants.js
@@ -1,6 +1,5 @@
 // @ts-check
 const { raw } = require('objection')
-const PokemonFilter = require('./Frontend')
 
 const LEVEL_CALC =
   'IFNULL(IF(cp_multiplier < 0.734, ROUND(58.35178527 * cp_multiplier * cp_multiplier - 2.838007664 * cp_multiplier + 0.8539209906), ROUND(171.0112688 * cp_multiplier - 95.20425243)), NULL)'
@@ -29,13 +28,10 @@ const MAD_KEY_MAP = /** @type {const} */ ({
   cp: 'cp',
 })
 
-const STANDARD = new PokemonFilter()
-
 module.exports = {
   AND_KEYS,
   IV_CALC,
   LEVEL_CALC,
   MAD_KEY_MAP,
-  STANDARD,
   BASE_KEYS,
 }
